requirements.txt
autopep8==1.6.0
click==8.1.3
Flask==2.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.8.0
python-crontab==2.6.0
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2
Werkzeug==2.1.2
zipp==3.8.1
